dom/tests/mochitest/pointerlock/file_constantXY.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - constantXY.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>
  <div id="div"></div>
  <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Confirm that screenX/Y and clientX/Y are constant when the mouse
       * is locked.
       */

      SimpleTest.waitForExplicitFinish();
      var pointer = navigator.mozPointer;
      var div;
      var lastScreenX, lastScreenY, lastClientX, lastClientY;

      var moveUnlocked = function (e) {
        is(pointer.isLocked, false, "Mouse should be unlocked");

        lastScreenX = e.screenX;
        lastScreenY = e.screenY;
        lastClientX = e.clientX;
        lastClientY = e.clientY;

        // Remove the current mousemove listener from "div"
        div.removeEventListener("mousemove", moveUnlocked, false);

        pointer.lock(div, function () {
          div.addEventListener("mousemove", moveLocked, false);

          // Move the mouse to a new location
          synthesizeMouse(div, 20, 20, {
            type: "mousemove"
          }, window);
        });
      };

      var moveLocked = function (e) {
        // Mouse should be locked here
        ok(pointer.isLocked, "Mouse should be locked");

        // Confirm that mouse coords are constant while locked
        is(e.clientX, lastClientX,
           "clientX should be equal to where the mouse was originaly locked");
        is(e.clientY, lastClientY,
           "clientY should be equal to where the mouse was originaly locked");
        is(e.screenX, lastScreenX,
           "screenX should be equal to where the mouse was originaly locked");
        is(e.screenY, lastScreenY,
           "screenY should be equal to where the mouse was originaly locked");

        document.mozCancelFullScreen();
      };

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreen && document.mozFullScreenElement === div) {
          // Moving the mouse when the pointer is not locked
          synthesizeMouse(div, 10, 10, {
            type: "mousemove"
          }, window);
        } else {
          SimpleTest.finish();
        }
      }, false);

      function start() {
        div = document.getElementById("div");
        div.addEventListener("mousemove", moveUnlocked, false);
        SimpleTest.waitForFocus(function() {
          div.mozRequestFullScreen();
        });
      }
  </script>
</body>
</html>
